Highlight the Tasks Assigned to Me!!!
I hate that this whole extension is the product of the complete
suckyness of TFS.  sigh...

This gives you a choice of colors for how you'd like your name
highlighted.

I tried to use the HTML5 <input type="color"/> element to give users a
full range of color choices, but as soon as you click on the Color
input, the Chrome popover disappears.  I looked into that, and there's
an open bug in the Chromium project about the issue.  Derp.

Anyhow... I suck at design, and layout.. if you want to make
improvements to how I added the controls to the popover... please do.

**Note: This commit is dependent on my previous pull request**