Add highlighting for work items labeled P1 and P2
I've wanted to do this for like 6 months, so i finally took the lazy
route and did it the easy route that doesn't support a ton of options.

Users have to create tasks that match with /^P[12][.: ].*/... it turns
RED for P1 and YELLOW for P2.

I had wanted to add a place for users to customize pattern matches and
colors, but meh.

Also, I was definitely too lazy to add a new option to enable/disable
this feature. Everyone gets it.  Let me know if that's not OK.